Add skeleton files
Added basic build configuration files and skeleton files for each part
of the engine that will be filled in the future. This includes
`core/``which is used for common functionality, `window/` to manage the
root window and widgets for the application, `opengl` for GL specific
components, and the same for `vulkan`.